use int64 for paymetn token quantity

+++ chaincodes/carbon/payment/virtual_token.go
@@ -16,8 +16,8 @@ const (
 // VirtualTokenWallet is a wallet that stores tokens corresponding to a
 // fiat currency. It is used to pay for the carbon credits.
 type VirtualTokenWallet struct {
-	OwnerID  string  `json:"owner"`
-	Quantity float64 `json:"quantity"`
+	OwnerID  string `json:"owner"`
+	Quantity int64  `json:"quantity"`
 }
 
 var _ state.WorldStateManager = (*VirtualTokenWallet)(nil)
